__m128[N] 的零成本抽象测试成功，接下来继续重构剩余的Batch类型

+++ tests_asm/math.cpp
@@ -10,63 +10,63 @@ namespace MyNamespace
     namespace KSIMD_DYN_INSTRUCTION
     {
         KSIMD_DYN_FUNC_ATTR void kernel(
-            const double* KSIMD_RESTRICT src,
-                  double* KSIMD_RESTRICT dst,
+            const float* KSIMD_RESTRICT src,
+                  float* KSIMD_RESTRICT dst,
             const size_t                 size
         ) noexcept
         {
             namespace ext = ksimd::ext::KSIMD_DYN_INSTRUCTION;
-            using f64 = KSIMD_DYN_BASE_OP(double);
+            using f32 = KSIMD_DYN_BASE_OP(float);
 
-            const f64::batch_t c10 = f64::set(0.123);
-            const f64::batch_t c9 = f64::set(-0.456);
-            const f64::batch_t c8 = f64::set(0.789);
-            const f64::batch_t c7 = f64::set(-0.101);
-            const f64::batch_t c6 = f64::set(0.234);
-            const f64::batch_t c5 = f64::set(-0.567);
-            const f64::batch_t c4 = f64::set(0.890);
-            const f64::batch_t c3 = f64::set(-1.234);
-            const f64::batch_t c2 = f64::set(0.555);
-            const f64::batch_t c1 = f64::set(1.999);
-            const f64::batch_t c0 = f64::set(-0.777);
+            const f32::batch_t c10 = f32::set(0.123);
+            const f32::batch_t c9 = f32::set(-0.456);
+            const f32::batch_t c8 = f32::set(0.789);
+            const f32::batch_t c7 = f32::set(-0.101);
+            const f32::batch_t c6 = f32::set(0.234);
+            const f32::batch_t c5 = f32::set(-0.567);
+            const f32::batch_t c4 = f32::set(0.890);
+            const f32::batch_t c3 = f32::set(-1.234);
+            const f32::batch_t c2 = f32::set(0.555);
+            const f32::batch_t c1 = f32::set(1.999);
+            const f32::batch_t c0 = f32::set(-0.777);
 
             size_t i = 0;
-            for (; i + f64::TotalLanes <= size; i += f64::TotalLanes)
+            for (; i + f32::TotalLanes <= size; i += f32::TotalLanes)
             {
-                f64::batch_t x = f64::load(src + i);
+                f32::batch_t x = f32::load(src + i);
 
-                f64::batch_t res = ext::vmath::clamp<f64>(c10, x, c9);
-                res = ext::vmath::lerp<f64>(res, x, c8);
-                res = ext::vmath::lerp<f64>(res, x, c7);
-                res = ext::vmath::lerp<f64>(res, x, c6);
-                res = ext::vmath::lerp<f64>(res, x, c5);
-                res = ext::vmath::clamp<f64>(res, x, c4);
-                res = ext::vmath::clamp<f64>(res, x, c3);
-                res = ext::vmath::clamp<f64>(res, x, c2);
-                res = ext::vmath::clamp<f64>(res, x, c1);
-                res = ext::vmath::clamp<f64>(res, x, c0);
+                f32::batch_t res = ext::vmath::clamp<f32>(c10, x, c9);
+                res = ext::vmath::lerp<f32>(res, x, c8);
+                res = ext::vmath::lerp<f32>(res, x, c7);
+                res = ext::vmath::lerp<f32>(res, x, c6);
+                res = ext::vmath::lerp<f32>(res, x, c5);
+                res = ext::vmath::clamp<f32>(res, x, c4);
+                res = ext::vmath::clamp<f32>(res, x, c3);
+                res = ext::vmath::clamp<f32>(res, x, c2);
+                res = ext::vmath::clamp<f32>(res, x, c1);
+                res = ext::vmath::clamp<f32>(res, x, c0);
 
-                f64::store(dst + i, res);
+                f32::store(dst + i, res);
             }
 
             // 尾处理
             if (const size_t tail = size - i; tail > 0)
             {
-                const f64::mask_t mask = f64::mask_from_lanes(static_cast<unsigned int>(tail));
-                f64::batch_t x = f64::mask_load(src + i, mask);
+                const f32::mask_t mask = f32::mask_from_lanes(static_cast<unsigned int>(tail));
+                f32::batch_t x = f32::mask_load(src + i, mask);
 
-                f64::batch_t res = ext::vmath::clamp<f64>(c10, x, c9);
-                res = ext::vmath::lerp<f64>(res, x, c8);
-                res = ext::vmath::lerp<f64>(res, x, c7);
-                res = ext::vmath::lerp<f64>(res, x, c6);
-                res = ext::vmath::lerp<f64>(res, x, c5);
-                res = ext::vmath::lerp<f64>(res, x, c4);
-                res = ext::vmath::clamp<f64>(res, x, c3);
-                res = ext::vmath::clamp<f64>(res, x, c2);
-                res = ext::vmath::clamp<f64>(res, x, c1);
-                res = ext::vmath::clamp<f64>(res, x, c0);
+                f32::batch_t res = ext::vmath::clamp<f32>(c10, x, c9);
+                res = ext::vmath::lerp<f32>(res, x, c8);
+                res = ext::vmath::lerp<f32>(res, x, c7);
+                res = ext::vmath::lerp<f32>(res, x, c6);
+                res = ext::vmath::lerp<f32>(res, x, c5);
+                res = ext::vmath::lerp<f32>(res, x, c4);
+                res = ext::vmath::clamp<f32>(res, x, c3);
+                res = ext::vmath::clamp<f32>(res, x, c2);
+                res = ext::vmath::clamp<f32>(res, x, c1);
+                res = ext::vmath::clamp<f32>(res, x, c0);
 
-                f64::mask_store(dst + i, res, mask);
+                f32::mask_store(dst + i, res, mask);
             }
         }
     } // namespace KSIMD_DYN_INSTRUCTION
@@ -77,8 +77,8 @@ namespace MyNamespace
 {
     KSIMD_DYN_DISPATCH_FUNC(kernel)
     void kernel(
-        const double* KSIMD_RESTRICT src,
-              double* KSIMD_RESTRICT dst,
+        const float* KSIMD_RESTRICT src,
+              float* KSIMD_RESTRICT dst,
         const size_t                 size
     ) noexcept
     {
